Redefine typing errors raised in NAST checks to NAST check errors
Summary:
A number of the 'typing' errors are actually nothing to do with typing. In a subsequent diff we will require that a typing environment be provided for evaluation so this diff removes those errors.

The errors we move here are all raised in NAST checks rather typing so we cannot (nor would we need to) provide a typing environment to evaluate them

Reviewed By: shayne-fletcher

Differential Revision: D45904391

fbshipit-source-id: 68209499534b498fe69a1d4743cbe39d93000130