refactoring: make typing_progress abstract
Summary:
I was bothered by how the typing_progress is updated and examined across multiple functions, and how we end up calling `failwith` as a result. It also makes updating code around it difficult.

In this diff, I make `typing_progress` abstract and put all the logic related to updating the progress and actually progressing through the `progress` in the enclosing new module TypingProgress.

Reviewed By: ljw1004

Differential Revision: D45608950

fbshipit-source-id: b9dafa1051d906805b706bc7ab63989589315e8a