Add array (vec) printer (including its elements), creation in test binary, Python unit tests
Summary:
Prints the contents of ArrayData[Vec]. Uses LLDB's SyntheticChildProvider API.

The next diff will add the ArrayData[Dict] case (right now, because of the way the synthetic child provider API works, it's not easy (or possible) to do have Dict and Keyset cases fall back to the default. So I print out a message saying to print them out in the raw format until it's implemented.

Reviewed By: oulgen

Differential Revision: D45964600

fbshipit-source-id: 335b49f75baf664a1dd3dcfe407fe01dd8954c6a