Use a correct widening cast for lowptr->ptr casts
Summary:
`SBValue::Cast` merely reinterprets bits, when we are extending a value from 4
bytes to 8 bytes we need to properly 0 extend (`SBValue::Cast` can put garbage in
the upper bits).

I originally went to make all casts in scripts use expressions, and got quite
close to getting it to work, but several issues made this difficult:
1) `SBValue`'s can't easily be used in expressions.  The `path` member has bugs
in it that make it unusable directly.  I wrote a fixup function that managed to
skirt around this by adding parenthesis to change operator precedence.
2) Evaluating expressions fully resolves them into values which break attempts
to then grab pointers to the newly created values.  There were a good number of
casses that relied on being able to rematerialize a pointer to a value that had
been dereferenced.
3) Expression based casts can't cast aggregate types (used in one place that
could have been fixed)

This also goes through and fixes up tests and funcVec logic in order to support
lowptr.

Reviewed By: mdko

Differential Revision: D45970044

fbshipit-source-id: 9c0e056b8fb810982a6aa3d7d5535d40a2504c3a