Handle JitResumeAddr properly
Summary: `JitResumeAddr` has a handler and an arg, both of which needs to be returned to `dispatchBB`. As both handler and arg are TCAs which are 4 bytes long, we can pack them into a 8 byte register. This enables us to return a `PcPair` in `rax` and `rdx` registers where `rdx` contains a packed `JitResumeAddr`.

Reviewed By: edwinsmith

Differential Revision: D45846106

fbshipit-source-id: ab62969116515466217f46b6a502516c12bcde4b