Improve treadmill crash reporting
Summary:
When dumping for crash, use a timed lock get instead of a try
lock. This gives any other process up to a second to release the
lock. If this times out, we'll proceed anyways as normal. Also, when
sending SIGABRT to another request, drop the lock first, since that
thread will try to grab it. Don't keep going, as the server is
crashing anyways.

Reviewed By: mdko

Differential Revision: D46006056

fbshipit-source-id: 47986e416c855f94a7f0b9af1b8f79d07c9fa261