Make matchers less noisy
Summary:
The current implementation prints the value of the thrift struct.
Especially when there are a few of them nested, it results in a lot of printed lines, mostly redundant.

Since gtest is already printing the value of the structs/union (we now have universal printing with `thrift/lib/cpp2/util/gtest/Printer.h`), it's now redundant to also print in the matcher

Reviewed By: vitaut

Differential Revision: D45770518

fbshipit-source-id: 4323aaf5ca59b6b7dcb8f972c4cd21424ef79423