make method invocation aware of module level traits
Summary:
During method invocation, module boundary checks are performed at method resolution time, eg. when performing `lookupMethodCtx`, relying on the caller information provided by a `MemberLookupContext`.

In absence of Module Level Traits, all the methods of a class belong to the same module of the class itself and a `MemberLookupContext` was implemented as:
```
 Either<const Class*, const StringData*> m_data;
```
preferring the class module information over the explicit module name information whenever both are available.

In presence of Module level traits, methods inlined in a class from a trait can belong to a different module than the class module.  This diff updates a `MemberLookupContext` so that it stores both the class _and_ explicit module name information, effectively making method invocation aware of module boundaries introduced by module level traits.

To ensure that correct module name information are always specified this diff removes the constructor below, which always infers the module name from the context class (which might be incorrect with module level traits):
```
explicit MemberLookupContext(const Class*); // used when we know the class is nonnull, or we are okay with module being null
```
The diff updates all the existing calls to the `MemberLookupContext(const Class*)` constructor to pass an explicit module name.

Reviewed By: oulgen

Differential Revision: D45569073

fbshipit-source-id: 93323871fc4abcac72fdb9030a87232eb686ad2b