Turn on implicit Sound Dynamic for more tests
Summary:
Apart from `test/typecheck` and `test/hover`, we're not running our tests under implicit pessimisation. Let's do this gradually by specifying `--enable-sound-dynamic-type` and `--everything-sdt` in `HH_FLAGS`. In this diff, a couple of issues are raised (and fixed), in `test/quickfixes`, and in `test/identify_symbol`.

The problems in `quickfixes` are due to `supportdyn` and `Tlike` in the types.

The problem, in `identify_symbol` is due to our implicit pessimisation adding an `as supportdyn<mixed>` bound even to transparent `type` definitions!

Differential Revision: D45992925

fbshipit-source-id: a719cc7cd7045c1edb594135d96b84c18dba4135