Make non-primitive cast account for like types
Summary:
Non-primitive cast lint warns when someone casts something to a non-primitive that'll result in an error at runtime.  This is done only for the cases where we know the runtime will occur.

With the release of true types, the subtyping check used to implement this lint started failing.

This diff relaxes the upper bound with like types to avoid false positives.

Reviewed By: andrewjkennedy

Differential Revision: D46041170

fbshipit-source-id: d4f8c79bbe0ea885adb220934b10a74493e86cb2