Make switch case scrutinee mismatch use disjointness
Summary:
With true types, the lint rule that detects when switch case and scrutinee can never match become noisy because in many cases a type (say an enum) `E` became `~E` and the rule was looking for strict subtyping relationships that were no longer satisfied.

This diff reduces the false positives by switching the implementation to use type disjointness relation.

Not only this got rid of the bug, but it also eliminated false positives that existed prior to the release of true types.

**N.b.** The test cases show funny results for the exhaustiveness check, but that's preexisting behaviour and this diff aims to address problems with case/scrutinee mismatch.

Reviewed By: vsiles

Differential Revision: D46038491

fbshipit-source-id: b7eb37d25f611545383d9c8d7b7037c6399a4d22