typo in sound-dynamic quick-fix test
Summary:
This is what the test looks like before this diff:
```
<<__SupportDynamicType>>
function main2(Wrap<float> $wf): void {
  (float)$wf;
}

<<__SupportDynamicType>>
function main2(vec<int> $_): void {
  (float)42.0;
}
```
Note that the two functions have the same name. The way typechecking works is it iterates over the list of toplevel names ["main2"; "main2"] and for each one, fetches the winning AST for that name from within the file, and typechecks its body.

This means that we end up typechecking the second function twice, and the first not at all. Which is weird.

Anyway, the test case is supposed to be about sound-dynamic, not about duplicate-name-behavior. I changed the test case accordingly.

Reviewed By: madgen

Differential Revision: D45882472

fbshipit-source-id: 9d610f9432ebacb05a8c7c560f1c7552583454d1