Hack support for module level traits
Summary:
With "module level traits" methods of public traits are considered part of the module they are defined in, instead of the module they are copied into.  This allows them to call internal top-level functions or methods of the module they were defined in, allowing developers to export internal functionality in a controlled way.  The module trait semantics is opt-in, and traits must be explicitly annotated with `<<__ModuleLevelTrait>>`.

HHVM enforcement is implemented in the stack that starts at D44025995.

This diff implements the "module level trait" semantics in the Hack module boundary checks.  It also turns `<<__ModuleLevelTrait>>` into an experimental feature, in `Unstable` status.

Reviewed By: jamesjwu

Differential Revision: D46015919

fbshipit-source-id: 28a5f136b995376a1674dd629427af5af051d4ac