Drop F14 prefetch()
Summary:
`prefetch()` was introduced in 2021 and since then it has found no application in our internal codebase.
The name is misleading, as this is actually a "prefetch more", to complement the one done in `prehash()`.
`prehash()` is also misleading, the prefetching operation should be independent, in fact virtually all the uses of `prehash()` in our codebase do not need prefetching.

So the plan would be
- Remove `prefetch()`, so we can confirm that it is not used anywhere and we can repurpose the name
- Split `prehash()` into `prehash()` and `prefetch()`, where `prehash()` only prehashes and `prefetch()` only prefetches the initial cache line, which is more intuitive.
- If a deeper prefetch is needed, it can be reintroduced not as a complement to `prefetch()`, but as a replacement (thus also prefetching the first cache line), or possibly as a template argument to `prefetch()` with tunable depth.

Reviewed By: terrelln

Differential Revision: D46018951

fbshipit-source-id: c13ddf6dedfd8e8ba616eddeb0ee0e6abbf89bd5