Don't set AttrInterceptable on builtins
Summary:
In order to be able to inline functions in sandbox mode, they cannot be renamable or interceptable. We are going to restrict which functions are renamable using a list in the config. Builtins are already not interceptable. We haven't looked into restricting these further.

This diff enforces the existing behavior of not allowing builtins to be interceptable. It also hooks up the RestrictRenameFunction runtime flag to enforce that builtins are not renamable, once we support that on the HHVM side and turn it on.

Reviewed By: aorenste

Differential Revision: D45883306

fbshipit-source-id: b7d978b42ee0e2fe95a5cfaa8a826627468e8d4b