Boost redundant/impossible type test signal
Summary:
Type tests via `is` and `as` can be redundant, e.g., `42 as arraykey` or impossible, e.g., `42 as bool`.

The existing linters weren't making use of the full subtyping and type disjointness algorithms significantly restricting the utility of the lints.

This diff boosts these linters using subtyping and type disjointness predicates.

It also eliminates some lint codes as they are just specific instances of subtyping or disjointness: 5611, 5626, 5630

Reviewed By: vsiles

Differential Revision: D45863928

fbshipit-source-id: 141d88eddcf841f136d9e1931dd86b833deea184