Prep hhvm for llvm15
Summary:
llvm15 seems to have an issue with uninitialized memory (which is UB anyway), so initializing it. I did a quick search and the only usage seems to be in checking the CFG, so this shouldn't impact perf afaik https://fburl.com/code/rephslc3.

Swayze with this change up the stack in D46079104 show overall win

Reviewed By: binliu19

Differential Revision: D45736726

fbshipit-source-id: be1c84057cf67364ec4cf69e7304d12b7ab62edd