Remove unused DeclResult::nopos_hash
Summary:
This was added to enable HHVM to ignore trivial whitespace changes, so
such changes to decls don't trigger recompiles. But it's dead code, and
optimizing unnecessary recompiles is going to be easier if we can start with
known-working code and then land optimizations that work end-to-end.

Also, if/when hackc starts reporting errors that contain decl positions (like typing),
this wouldn't be a valid optimization anyway.

Reviewed By: aorenste

Differential Revision: D46085602

fbshipit-source-id: d9b01ecc7824cca19bc5a63a884ee54032ce5adc