Rename decl-parsing APIs that are overly heavyweight
Summary:
hackc::direct_decl_parse() returns a DeclResult containing parsed
decls as a walkable data structure *and* as a serialized blob.

For most callers the extra serialization step is not necessary and we
should refactor the API to avoid wasted work. As a first step, rename
things to make it clear when we're using the heavy duty versions.

Reviewed By: aorenste

Differential Revision: D46109358

fbshipit-source-id: 518da25428aa33ea8604cb6808bd03bceaa08964