win32: handle overflow events
Summary:
According to Microsoft's documentation, the kernel may overflow its own buffers
if user space doesn't read file notification fast enough. In that case, the
documentation is not clear, but a combinaison of 2 things can be done:
 - An ERROR_NOTIFY_ENUM_DIR will be returned
 - The written bytes value will be set to 0

From reading the code, it appears that ERROR_NOTIFY_ENUM_DIR is correctly
used[0], but the code never test for the written bytes value. Thus in the case
where Windows only notify Watchman of an overflow via the written bytes,
Watchman will not detect the overflow and miss events!

To better understand the error, I've simply added a sleep(20) right before
calling ReadDirectoryChangesW to force the buffer to overflow and copied a
large amount of files in a watched directory. Watching the directory with
`watchman-wait . -m 0` showed that no files were detected as written and that
GetOverlappedResult was setting the written bytes to 0.

Reviewed By: mshroyer

Differential Revision: D46122162

fbshipit-source-id: d6a0b65e51283ca7bb8ab160b84d3d034daebbaf