Account for supportdyn
Summary:
Predicate that checks if a function/method is memoized looks for a `Tfun` and if it cannot find one, it concludes that it is not memoized.

This broke when true types were released because function types are now wrapped around with `supportdyn<...>`. Hence, the predicate was nearly always returning false.

This diff fixes it by looking inside the `supportdyn<...>` wrapper.

This diff also adds `--implicit-pess` flag to the test directory so that the analysis is tested with true types turned on.

Differential Revision: D46147360

fbshipit-source-id: 08bb6c346a5b86d678516db63959a68682f22c1e