Remove unused internal OSS build artifacts
Summary:
These files used to be used for building OSS internally. We are no longer supporting it. Notice that they are under fb directory, so they are not shipped to OSS.

The only file not under fb directory is `hphp/test/cmake_builds_excluded_tests` which is not read anywhere else except for under this fb directory.

Reviewed By: edwinsmith

Differential Revision: D46159401

fbshipit-source-id: 4b5c069c8e9878b482047f98bfd946a5951fd18e