Add visited files tracking API
Summary:
This API is a prerequisite for further improvements to autoloading.
While very similar to get_included_files this API is only collecting files once it was turned on. Once turned on it should include a superset of get_included_files.
In the future the results of this API may vary, but would always be a superset of what get_included_files would have returned from the moment it is turned on.

Reviewed By: paulbiss

Differential Revision: D46080341

fbshipit-source-id: b263eed941f9e2e268673034b5d9d966d97aec79