get_full_pos_from_decl
Summary:
For duplicate names, the existing paradigm is: after a watchman event, as part of our processing of disk-changes, we update the reverse naming table and in the process of doing so we report duplicate names.

I wish to change this so that duplicate names are reported by the typechecking code, e.g. Typing_toplevel.nast_to_tast and Typing_check_job.calc_errors_and_tast, so it can be shared by distc and clientIdeDaemon and more easily by hh_single_type_check. (Currently duplicate names are reported in hh_server's main loop, in serverTypeCheck.ml). That happens in the next diff; this diff provides the groundwork.

In the course of this work I discovered that existing APIs are kind of surprising in the context of duplicate names, e.g. `Decl_provider.get_typedef` will happily return a typedef even if it's not the winner. (hence, how can a different piece of code typecheck against?) Also, the existing APIs are really convoluted -- e.g. you are given a name, then
1. consult Naming_provider to get the filename
2. parse the filename into an AST
3. walk that filename to find which syntax definition corresponds to the name you were given
4. obtain from the AST its canonical capitalization
5. query the decl service to get the decl

I wish to streamline all of this. It's enough to simply get the position-full decl and read its position directly! Indeed that's what we currently do in Typing_class.class_def. However, what we currently do in Typing_class.class_def is incorrect (per the above) in the presence of duplicate names or non-canonical capitalization.

What I'm doing in this diff is creating a well-tested, robust, and high-level API for telling whether a given definition is the winning definition, as cheaply as possible.
```
val is_this_def_the_winner :
  Provider_context.t -> FileInfo.name_type -> Ast_defs.id -> winner
```
I deliberately made it as high-level as possible because I wish to replace a lot of Naming_provider and Decl_provider with something more careful and correct and tightly bound. I also wanted to make an API that would handle things easily if we decide to switch to position-free decls. Explanation of how we'll do this is in the docblock.

Differential Revision: D46073212

fbshipit-source-id: 612c459967248edc284c0227c5013cac62a0bfea