3/5 Typing_check_job.check_* takes ~full_ast, not fn+name
Summary:
Overview of this stack of five [type-ast] diffs is fully explained in D45834281. Summary:
* Currently the typechecker takes a list of top-level names to typecheck, then iterates over that list of names, fetches the winning definition of that name in the file, and typechecks it.
* Thanks to this stack, the typechecker will instead iterate over the list of definitions in the file and typecheck them.

---------

This is the principal diff in the stack. It changes Typing_check_job, e.g.
```
(* BEFORE *)
val type_fun :
  Provider_context.t -> Relative_path.t -> string -> Tast.def list option

(* AFTER *)
val type_fun :
  Provider_context.t -> full_ast:Nast.fun_def -> Tast.def list option
```
This diff also changes the callers to iterate over their ASTs and send them to typing_check_job, rather than iterating over names.

We have two kinds of ASTs -- *full* ASTs which include the bodies of methods, and "quick-mode" ASTs which don't. It's naturally important that the caller provide full asts to these typechecking functions. There's nothing in the type-system to enforce this. What I've done for now is make all the functions take a named parameter `~(full_ast : Nast.*)` so it's very clear at the callsite what's needed. Previously this wasn't needed because Typing_check_job was responsible for looking up the full ast from Ast_provider.

I think the new behavior is better because it's just weird that the caller should obtain an AST to get names and leave it in cache, and then Typing_check_job should obtain the AST from a cache a second time. This new design will allow us to skip the cache entirely.

It's still unpleasant to have two forms of AST. I think that the named parameter`~(full_ast : Nast.*)` is a proportionate and sufficient mitigation for now. But I think the proper fix is to eliminate quick-mode ASTs altogether. Now that we have the direct-decl-parser, any time we do quick-mode is probably a mistake. I started a separate diff-stack to see if I can get rid of them. D45892426

Differential Revision: D45836982

fbshipit-source-id: 14299dc4eda48dcd7806455fe201c88717a590b7