get_full_pos_by_parsing_file
Summary:
This diff (1) is a no-op diff to rename
```
Naming_provider.get_full_pos --> get_full_pos_by_parsing
```
(2) eliminates some special-case code for the now-defunct part of the rearchitecture.

I'm writing this diff because the act of fetching and parsing an AST solely in order to be able to get the position of a decl (which we solely do in order to tell whether we have in hand a duplicate name) is kind of heavyweight. Elsewhere in Typing_class we use another nicer technique, which is to get the position-full decl and extract position from that.

At the very least, we should have function names that make it clear which of the two techniques we're using! Hence this rename,

Differential Revision: D46052575

fbshipit-source-id: c631ef676ed972f39cce39c16b4ae91f05614d3e