Make handling more resilient for Eval.UnixServerRunPSPInBackground
Summary: When the unix server is unavailable we can fallback to local execution, however in this case the background process will complete execution and then exit without sending back an exit code. The logic in `moveToBackground` should handle socket disconnections (read returns 0) by waiting for the background process to exit and then exiting with the same code.

Reviewed By: ricklavoie

Differential Revision: D46041511

fbshipit-source-id: 282686743dd610053fe9729b048836ee7f2459c1