5/7 saved-state with dupes?
Summary:
[Overview: I'm changing it so duplicate-names are no longer reported during hh_server's main loop when it calls Naming_global; instead they'll be reported along with every other parse+typing error during Typing_toplevel]

The current code makes a pretense of handling saved-states with naming errors, but it doesn't actually work reliably. Check this out:
```
$ touch .hhconfig
$ printf "<?hh\nclass Car {}\n" > a.php
$ printf "<?hh\ntype car = int;\n" > b.php
$ hh --no-load
b.php: Name already bound: `car`

$ hh stop
$ ~/www/scripts/hack_for_hiphop/hh_ss.sh # generate
$ hh stop
$ ~/www/scripts/hack_for_hiphop/hh_ss.sh # load from ss
0 errors found
$ touch a.php
$ touch b.php
$ hh
0 errors found
```
Here's an example with three:
```
$ hh stop
$ printf "<?hh\nclass Car {}\n" > a.php
$ printf "<?hh\ntype car = int;\n" > b.php
$ printf "<?hh\ntype CAR = int;\n" > c.php
$ hh --no-load
b.php: Name already bound: `car`
c.php: Name already bound: `CAR`

$ hh stop
$ rm -fr /tmp/hh_ss
$ ~/www/scripts/hack_for_hiphop/hh_ss.sh # generate
$ hh stop
$ ~/www/scripts/hack_for_hiphop/hh_ss.sh # load from ss
c.php: Name already bound: `CAR`
$ rm c.php
$ hh
0 errors found
$ touch a.php
$ touch b.php
$ hh
0 errors found
```

The design of this feature is simply wrong. Hh_server requires the invariant that `env.failed_files` contains all files involved in a naming conflict, both the winner and the loser(s). But when we load from saved state, we only perform naming on the loser(s). Without the invariant, the examples I gave are obvious consequences.
```
let naming_files =
    List.fold
      ~init:Relative_path.Set.empty
      ~f:Relative_path.Set.union
      [
        naming_error_files;
```

Since this logic in serverLazyInit.ml only works for some cases but not others of saved-state-load with duplicate-name errors, I don't think it's worth having. I'd rather remove it, and start properly from a clean slate elsewhere if we deem it important. (News flash: we'll never deem it important! The only reason that saved-states-with-errors exist is to generate shadow saved states for the new version of hack even if the new version of hack introduces new errors, but duplicate-name errors aren't the kind of "new error" it will ever introduce).

If we wanted to fix it up, there are more principled approaches:
1. deny saving a saved-state if there exist duplicate names, i.e. if `env.failed_naming` is non-empty. This is principled because it truly reflects the fact that our current on-disk saved-state format is not expressive enough to reflect duplicate name situations.
   1. (I didn't do this because there are so many tools that save state, and as I started to thread `env.failed_naming` through to all of them it became quite non-obvious how to proceed in some. Not a straightforward change.)
2. Invent a modified saved-state format which can express duplicate names. bobrenjc93 already invented such a format for the re-architecture, with his "overflow table". Once I move on to work on T146523780 I'll adopt this.

Reviewed By: CatherineGasnier

Differential Revision: D46121543

fbshipit-source-id: aa30bbe26ca350cfd462ee1e9a8adc7dc4920292