Don't report 'not dynamically callable' under implicit pessimisation
Summary:
Under implicit pessimisation, there should be no "not dynamically callable" failures. Any failures of this sort indicate bugs in the sound dynamic implementation and is not useful to the user.

To mitigate this, this diff only reports these errors when implicit pessimisation is turned off (either by setting a flag or indirectly by adding `__NoAutoDynamic`).

Reviewed By: andrewjkennedy

Differential Revision: D46193259

fbshipit-source-id: 341292daeabfcc9fafed6ae36f342bccc101d94f