decls_to_facts_json()
Summary:
In the critical path of parsing decls for factsdb, we do several conversions:
1. direct_decl_parse_and_serialize() - parse decls from source text
2. decls_to_facts() - convert to facts::Facts
3. decls_to_facts() - convert facts::Facts to compiler_ffi::Facts
4. facts_to_json() - convert compiler_ffi::Facts to facts::Facts
5. facts_to_json() - convert facts::Facts to JSON via serde_json
6. parse_json() - convert JSON to folly::dynamic
7. make_file_facts() - convert folly::dynamic to HPHP::Facts::FileFacts
8. update SymbolMap & DB from FileFacts

This diff deletes step 3 and 4 by replacing facts_to_json() with decls_to_facts_json(),
which avoids converting in and out of the intermediate compiler_ffi::Facts representation.

why do we have both compiler_ffi::Facts and facts::Facts? the types declared in
the cxx bridge are automatically available to both Rust and C++, but using them
outside the context of hhvm would create a dependency on the bridge.
The best solution is stop using facts::Facts everywhere outside the context of hhvm,
but that will take more archeology.

Reviewed By: nt591

Differential Revision: D46149159

fbshipit-source-id: a1db9f99d6957d5fcff9ab07e20f2353c8a02c1e