Enforce deployment boundaries for top level functions
Summary:
Raise an exception when calling a toplevel function that is not within a module included in the current active deployment. This diff handles calling
- directly with `foo()`
- dynamically with `$f = "foo"; $f()`
- via function pointers `$f = foo<>; $f()`
- via closures `$f = () ==> { return foo(); }; $f()`

Note that enforcement on class method calls will be implemented after enforcement on classes is implemented. I'm happy to address any other cases that this diff is missing

Reviewed By: oulgen

Differential Revision: D45644101

fbshipit-source-id: 6a13361e161b6fb7c2b86e29548491d3068ba78d