(Conservatively) Check if Case Type Variants are Disjoint
Summary:
# Context
Case Types currently allow any union to be assigned in its definition. We want to limit the possible unions that can be written to include only "well-behaved" unions that can be decomposed by examining only the shallow tag of a value.

# High-Level Approach
Every value at runtime is associated with a particular tag. The set of potential tags are listed [here](https://www.internalfb.com/code/fbsource/fbcode/hphp/runtime/base/datatype.h) in HHVM. Since types describe sets of values, we can map each type to a set of tags that it can contain. If the set of tags for two different `T1` and `T2` are disjoint, then you can check the tag of a value at runtime to determine whether it has type `T1` or `T2`. If all the variant types of a case type are disjoint then we can always decompose a case types into its variant types. If the variants overlap, then we may not be able to decompose to a singular type. This diff is an initial, very conservative version of this check that examines the variant types of a case type and determines if they are disjoint. It is conservative in a few regards:
1. If any of the variant types overlap at all, meaning they contain values whose shallow tags may overlap we produce an error. A less conservative approach could allow for some overlapping, i.e. `case type CT = arraykey | num`, which overlap for the `int` data type yet both map any values with that tag to be of type `int`
2. Any object type is considered to overlap, thus even if two classes are final and do not extend each other we will still disallow a case type containing them. A less conservative approach would consider the hierarchy of a class when determining if variant types overlap
3. The function type is assumed to contain the same tags as `nonnull`. There are some tags included in this set that I'm 99% sure cannot be used as a callable type (i.e. `bool`), but I elect for the more conservative interpretation to begin with.

# Implementation

I added a new file `Typing_case_types`. This contains a module `DataType` that contains another module `Tag`.

`DataType.Tag.t` is a variant type that matches with the set of data types in [HHVM](https://www.internalfb.com/code/fbsource/fbcode/hphp/runtime/base/datatype.h). This is a subset because some of the data types are not observable at the language level and are only present to allow for performance optimizations in the runtime.

`DataType.fromTy` maps a `locl_ty` to `DataType.t` which is a set of `DataType.Tag.t`. This mapping is mostly straightforward. There are a few types that need to be specially handled, particularly:
* The array types since they have corresponding tags, but are represented as objects in our hhi files.
* There are special interfaces that contain instances whose tags are no objects (i.e. XHPChild)

For the special interfaces care was taken not to duplicate logic, by determining what tags are present in these interfaces by utilizing subtype tests. See `DataType.special_interface_to_datatypes` for details.

The `DataType` module also contains a small algebra supporting union, intersect and set difference for sets of tags. This is used in the mapping of certain types to datatypes, such as `nonnull` which can be defined as `DataType(mixed) - DataType(null)`.

To determine overlapping, for each variant type of a case type we determine the set of tags it maps to. We then translate this to a map from tag to list of types. If this mapping contains a list whose length is greater than one, it means there is an overlap and we report an error. The checks are performed in `Typing_typedef` which contains the well-formedness checks performed on all typedefs (aliases, newtypes, casetypes).

# Improvements
Besides being less conservative in the analysis, the error messages could be improved. Specifically more care could be taken to explain the reason a certain type is associated with a particular tag. I suggest we land this anyways and wait until we can get feedback from users to determine what information would be useful to include in the error message.

Reviewed By: vsiles

Differential Revision: D45372775

fbshipit-source-id: 5750b739d204115d3d4c1601c5eb4e324dc180df