Set dwp file as a resource, so lldb can find debugging information in the unit tests
Summary:
D46199492 allowed the .dwp file to be created (so I could open up the test binary manually in lldb).

This diff makes it visible to the unit tests by adding it as a resource (and renaming it in the process, to `hhvm_types.dwp` so lldb will find it easily).

Reviewed By: jtwarren

Differential Revision: D46207207

fbshipit-source-id: 64a1c3aa09a3e0029bfb263b1155baf46e042825