Remove the JSONC parser (#9365)
Summary:
The JSONC parser is gated by USE_JSONC in CMake,
which is disabled by default and thus it wasn't
build in a long time.

As the JSONC parser seems unused for most users
of HHVM, it would make sense to only keep the
stricter built-in parser.

Pull Request resolved: https://github.com/facebook/hhvm/pull/9365

Test Plan: it builds

Reviewed By: Wilfred

Differential Revision: D45869197

Pulled By: paulbiss

fbshipit-source-id: aa8fe49f599d138f25a24a00fe8044a9cd16ef89