Fix local type resolution
Summary:
Recently(ish) "enhanced" type resolution allows referring to nonexistent types that may be generated at some later translation stage from the same program. This is extremely problematic for multiple reasons:

* It is confusing to users because it allows referring to entities that do not exist in the IDL and enables conflicts with user entities which we have always tried to avoid.
* Compilation process is very fragile because unresolved types have to be handled at every stage. As a result we have recurrent crashes and bad diagnostics both in the compiler and in the tooling. A lot of time has been spent on partial workarounds, debugging and responding to users.
* It makes impossible to separate generated and non-generated entities in the target code which will negatively affect build times for users that don't rely on the feature which is most Thrift users. For example, patch for a simple struct generates 1.5x times more code than the struct itself.
* There can be potentially problematic cyclic dependencies between generated and non-generated entities the effects of which are unclear.

Considering that there hasn't been general consensus that we want this model (it was added without design discussions as a byproduct of patch) and there are open design questions, revert to the previous status quo of resolving local types at the AST construction phase. This should solve most of the bugs observed by users and the rest can be fixed by applying a similar technique to included programs (to be done as a follow up).

If we want to allow referring to generated entities within the same program this needs further design work to address outstanding issues, ideally backed by experience from other languages.

This doesn't fully address the first (user confusion) problem but significantly constrains the damage from the new type resolution while we consider design options.

As a nice side effect this eliminates a few cascading diagnostics and an inconsistency in reporting unresolved annotations which were previously reported as errors in constants.

Reviewed By: iahs

Differential Revision: D46224310

fbshipit-source-id: 211e40dd0b8142471c32bc3feb7c490054297871