Use Relative_path.t instead of string in Rename
Summary:
We should be using "string" for filenames only at the untrusted periphery (e.g. CLI arguments, LSP messages), and convert them to Relative_path.t for use everywhere else.

I got bitten by this because the code in serverRename started with a Relative_path.t, did `suffix` on it, and then passed it to the rename engine which did `Relative_path.create_detect_prefix`. This naturally broke unless the CWD was identical to root!

The solution is that the rename engine should have accepted Relative_path.t, not string. That's what this diff does.

(Along the way I added some names to parameters where it was too hard to tell what they did otherwise).

Reviewed By: nt591

Differential Revision: D46239349

fbshipit-source-id: b28f9031822264e0591c64ce32069617445c0fd7