shellout pass root explicitly
Summary:
I'm working on porting the tests. The tests already set up CWD to be the repo dir, so when clientLsp shells out to hh then it's heuristic should be able to walk up the hierarchy and find the same .hhconfig.

But for some reason it wasn't happening for me! So, with this diff, I'm passing the root explicitly.

In any case, I kind of prefer having things as explicit as possible.

Reviewed By: nt591

Differential Revision: D46226570

fbshipit-source-id: be778ea9fe76525e12dcaa7a3e2af9106a099c93