rename clientIdeDaemon localvar needs dollar
Summary:
When renaming a local variable, if the user places their caret over `$fo|o` and presses F2 and renames it to `bar`, then we should replace with `$bar` i.e. implicitly insert the dollar sign.

This is already done by `go_ide` but not by `go_for_localvar` which is what ClientIdeDaemon uses. This diffs adds it there.

Reviewed By: nt591

Differential Revision: D46241670

fbshipit-source-id: 42dff2e46c1436435fc32c09e80ac624ea2b1473